% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbod.R
\name{gbod}
\alias{gbod}
\title{gbod}
\usage{
gbod(df, n_partition = 10, outlier_percent = 1)
}
\arguments{
\item{df}{: Input Dataframe, contains only the columns to be analyzed and no missing values}

\item{n_partition}{: integer, decide how many parts will be separated on each attributes}

\item{outlier_percent}{: number, the outlier percentage of the total records.}
}
\value{
dataframe with additional column: label, 0 means normal, 1 means outlier
}
\description{
Grid-based anomaly detection base algorithm
}
\examples{
gbod(df, n_partition=10, outlier_percent=1)
}
