% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbod.R
\name{gbod_score}
\alias{gbod_score}
\title{gbod score}
\usage{
gbod_score(df, partition_range = c(5, 8), outlier_percent = 1)
}
\arguments{
\item{df}{: input dataframe without missing values}

\item{partition_range}{: vector with two elements, a range of partitions, for example c(5,8)}

\item{outlier_percent}{: number}
}
\value{
dataframe with different partition result and score
}
\description{
calculate the outlier score for each record
}
\examples{
gbod_score(df, partition_range=c(5,8), outlier_percent=1)
}
