% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbod.R
\name{normalit}
\alias{normalit}
\title{normalization and scale}
\usage{
normalit(m, n_partition)
}
\arguments{
\item{m}{a vector to be normalized}

\item{n_partition}{integer, the factor to be scaled}
}
\value{
the normalized and scaled vector
}
\description{
normalizate a vector and scale it
}
