% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbod.R
\name{special_floor}
\alias{special_floor}
\title{round integer}
\usage{
special_floor(m, n_partition)
}
\arguments{
\item{m}{a vector to be rounded}

\item{n_partition}{integer, the factor to be scaled}
}
\value{
the rounded integer vector
}
\description{
round down integer function special for gbod which ignore the partition
}
